requirements.txt
aniso8601==10.0.1
annotated-types==0.7.0
attrs==25.3.0
blinker==1.9.0
click==8.2.1
Flask==3.1.1
flask-restx==1.3.0
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
MarkupSafe==3.0.2
mysql-connector-python==9.4.0
pytz==2025.2
referencing==0.36.2
rpds-py==0.26.0
typing-inspection==0.4.1
typing_extensions==4.14.1
Werkzeug==3.1.3
